完善同上 主键从numeric调整为int

+++ Core/Database/Logsystem.cs
@@ -14,7 +14,7 @@ namespace Database
     
     public partial class Logsystem
     {
-        public decimal ID { get; set; }
+        public int ID { get; set; }
         public Nullable<System.DateTime> Logtime { get; set; }
         public string Ident { get; set; }
         public string Priority { get; set; }
